Revert "Track prize updates"
