+++ app/prizes/PathPrize.tsx
@@ -84,7 +84,7 @@ const PathPrize: React.FC<PathPrizeProps> = ({
     }
 
     const scaleFinal = {
-        xs: 0.55,
+        xs: 0.6,
         sm: 0.51,
         md: 0.75,
         lg: 0.8,
@@ -106,65 +106,15 @@ const PathPrize: React.FC<PathPrizeProps> = ({
                 flexDirection: "column",
                 alignItems: "center",
                 width: size,
-                maxWidth: { xs: "75vw", sm: "unset" },
                 height: "auto",
                 position: "relative"
             }}
         >
-            {/* Mobile: flat title (≤ 600px) */}
-            <Box
-                sx={{
-                    display: "none",
-                    "@media (max-width: 600px)": {
-                        display: "flex",
-                        alignItems: "center",
-                        justifyContent: "center",
-                        gap: "6px",
-                        mb: 1
-                    }
-                }}
-            >
-                <Typography
-                    sx={{
-                        color: "#fff",
-                        fontFamily: "Tsukimi Rounded",
-                        fontWeight: 700,
-                        fontSize: "18px",
-                        textAlign: "center"
-                    }}
-                >
-                    {topText}
-                </Typography>
-                {showHelpIcon && helpTooltip && (
-                    <Tooltip
-                        title={
-                            <Typography
-                                sx={{
-                                    fontFamily: "Montserrat",
-                                    fontSize: "16px"
-                                }}
-                            >
-                                {helpTooltip}
-                            </Typography>
-                        }
-                    >
-                        <Box
-                            component="img"
-                            src="/prizes/path_prizes/question.svg"
-                            alt="help"
-                            sx={{ width: 20, height: 20 }}
-                        />
-                    </Tooltip>
-                )}
-            </Box>
-
             {/* 1. VISUAL CONTAINER (The Path/SVG) */}
             <Box
                 sx={{
                     width: size,
-                    maxWidth: { xs: "75vw", sm: "unset" },
                     height: size,
-                    maxHeight: { xs: "75vw", sm: "unset" },
                     backgroundImage: `url("${backgroundSrc}")`,
                     backgroundRepeat: "no-repeat",
                     backgroundSize: "contain",
@@ -178,152 +128,162 @@ const PathPrize: React.FC<PathPrizeProps> = ({
                     transformOrigin: "center"
                 }}
             >
-                <Box
-                    sx={{
+                <svg
+                    width="100%"
+                    height="100%"
+                    viewBox="0 0 400 400"
+                    style={{
                         position: "absolute",
                         inset: 0,
-                        "@media (max-width: 600px)": {
-                            display: "none"
-                        }
+                        marginTop: `${centerOffsetY}px`,
+                        marginLeft: `${centerOffsetX}px`,
+                        pointerEvents: "none"
                     }}
                 >
-                    <svg
-                        width="100%"
-                        height="100%"
-                        viewBox="0 0 400 400"
-                        style={{
-                            marginTop: `${centerOffsetY}px`,
-                            marginLeft: `${centerOffsetX}px`,
-                            pointerEvents: "none"
-                        }}
-                    >
-                        <circle
-                            cx="200"
-                            cy="200"
-                            r={radius * 1.2}
-                            fill="transparent"
-                            stroke="transparent"
-                            style={{ pointerEvents: "all" }}
-                            onMouseEnter={() => setHovered(true)}
-                            onMouseLeave={() => setHovered(false)}
-                        />
-                        <defs>
-                            {shouldGradient && (
-                                <linearGradient
-                                    id={topGradId}
-                                    x1="0%"
-                                    y1="0%"
-                                    x2="100%"
-                                    y2="0%"
-                                >
-                                    <stop
-                                        offset="0%"
-                                        stopColor={topGradient!.from}
-                                    />
-                                    <stop
-                                        offset="50%"
-                                        stopColor={
-                                            topGradient!.mid ?? topGradient!.to
-                                        }
-                                    />
-                                    <stop
-                                        offset="100%"
-                                        stopColor={topGradient!.to}
-                                    />
-                                </linearGradient>
-                            )}
+                    <circle
+                        cx="200"
+                        cy="200"
+                        r={radius * 1.2}
+                        fill="transparent"
+                        stroke="transparent"
+                        style={{ pointerEvents: "all" }}
+                        onMouseEnter={() => setHovered(true)}
+                        onMouseLeave={() => setHovered(false)}
+                    />
+                    <defs>
+                        {shouldGradient && (
+                            <linearGradient
+                                id={topGradId}
+                                x1="0%"
+                                y1="0%"
+                                x2="100%"
+                                y2="0%"
+                            >
+                                <stop
+                                    offset="0%"
+                                    stopColor={topGradient!.from}
+                                />
+                                <stop
+                                    offset="50%"
+                                    stopColor={
+                                        topGradient!.mid ?? topGradient!.to
+                                    }
+                                />
+                                <stop
+                                    offset="100%"
+                                    stopColor={topGradient!.to}
+                                />
+                            </linearGradient>
+                        )}
 
-                            <path
-                                id={topArcId}
-                                d={`
+                        <path
+                            id={topArcId}
+                            d={`
               M 200,200
               m -${radius},0
               a ${radius},${radius} 0 0 1 ${radius * 2},0
             `}
-                            />
+                        />
 
-                            <path
-                                id={bottomArcId}
-                                d={`
+                        <path
+                            id={bottomArcId}
+                            d={`
               M 200,200
               m -${radius},0
               a ${radius},${radius} 0 0 0 ${radius * 2},0
             `}
-                            />
-                            {secondText && (
-                                <path
-                                    id={secondId}
-                                    d={`
+                        />
+                        {secondText && (
+                            <path
+                                id={secondId}
+                                d={`
               M 200,200
               m -${radius - 25},0
               a ${radius - 25},${radius - 25} 0 0 0 ${radius * 2 - 50},0
             `}
-                                />
-                            )}
-                        </defs>
+                            />
+                        )}
+                    </defs>
 
-                        {/* top text */}
-                        <text
-                            fontFamily="Tsukimi Rounded"
-                            fontSize={topTextSize}
-                            fontWeight="700"
-                            textAnchor="middle"
+                    {/* top text */}
+                    <text
+                        fontFamily="Tsukimi Rounded"
+                        fontSize={topTextSize}
+                        fontWeight="700"
+                        textAnchor="middle"
+                    >
+                        <textPath
+                            href={`#${topArcId}`}
+                            startOffset={`${topTextOffset}%`}
                         >
-                            <textPath
-                                href={`#${topArcId}`}
-                                startOffset={`${topTextOffset}%`}
-                            >
-                                {shouldGradient ? (
-                                    <>
-                                        <tspan fill="#fff">{topBefore}</tspan>
-                                        <tspan fill={`url(#${topGradId})`}>
-                                            {topMid}
-                                        </tspan>
-                                        <tspan fill="#fff">{topAfter}</tspan>
-                                    </>
-                                ) : (
-                                    <tspan fill="#fff">{topText}</tspan>
-                                )}
-                            </textPath>
-                        </text>
+                            {shouldGradient ? (
+                                <>
+                                    <tspan fill="#fff">{topBefore}</tspan>
+                                    <tspan fill={`url(#${topGradId})`}>
+                                        {topMid}
+                                    </tspan>
+                                    <tspan fill="#fff">{topAfter}</tspan>
+                                </>
+                            ) : (
+                                <tspan fill="#fff">{topText}</tspan>
+                            )}
+                        </textPath>
+                    </text>
 
-                        {showHelpIcon && helpTooltip && (
-                            <Tooltip
-                                title={
-                                    <Typography
-                                        sx={{
-                                            fontFamily: "Montserrat",
-                                            fontSize: "16px"
-                                        }}
-                                    >
-                                        {helpTooltip}
-                                    </Typography>
-                                }
-                            >
-                                <g
-                                    style={{ pointerEvents: "all" }}
-                                    onMouseEnter={() => setHovered(true)}
-                                    onMouseLeave={() => setHovered(false)}
-                                    transform={`
+                    {showHelpIcon && helpTooltip && (
+                        <Tooltip
+                            title={
+                                <Typography
+                                    sx={{
+                                        fontFamily: "Montserrat",
+                                        fontSize: "16px"
+                                    }}
+                                >
+                                    {helpTooltip}
+                                </Typography>
+                            }
+                        >
+                            <g
+                                style={{ pointerEvents: "all" }}
+                                onMouseEnter={() => setHovered(true)}
+                                onMouseLeave={() => setHovered(false)}
+                                transform={`
         translate(${200 + radius * 1.06 * Math.cos((helpAngleDeg * Math.PI) / 180)},
                     ${200 + radius * 1.06 * Math.sin((helpAngleDeg * Math.PI) / 180)})
         rotate(${helpRotationDeg})
         translate(${-helpSize / 2}, ${-helpSize / 2})
         `}
-                                >
-                                    <title>{helpTooltip}</title>
+                            >
+                                <title>{helpTooltip}</title>
 
-                                    <image
-                                        href="/prizes/path_prizes/question.svg"
-                                        width={helpSize}
-                                        height={helpSize}
-                                        preserveAspectRatio="xMidYMid meet"
-                                    />
-                                </g>
-                            </Tooltip>
-                        )}
+                                <image
+                                    href="/prizes/path_prizes/question.svg"
+                                    width={helpSize}
+                                    height={helpSize}
+                                    preserveAspectRatio="xMidYMid meet"
+                                />
+                            </g>
+                        </Tooltip>
+                    )}
 
-                        {/* bottom text */}
+                    {/* bottom text */}
+                    <text
+                        fill="#fff"
+                        fontFamily="Montserrat"
+                        fontSize={bottomTextSize}
+                        fontWeight="600"
+                        textAnchor="middle"
+                        style={{ letterSpacing: bottomLetterSpacing }}
+                    >
+                        <textPath
+                            href={`#${bottomArcId}`}
+                            startOffset={`${bottomTextOffset}%`}
+                        >
+                            {bottomText}
+                        </textPath>
+                    </text>
+
+                    {secondText && (
                         <text
                             fill="#fff"
                             fontFamily="Montserrat"
@@ -333,67 +293,14 @@ const PathPrize: React.FC<PathPrizeProps> = ({
                             style={{ letterSpacing: bottomLetterSpacing }}
                         >
                             <textPath
-                                href={`#${bottomArcId}`}
+                                href={`#${secondId}`}
                                 startOffset={`${bottomTextOffset}%`}
                             >
-                                {bottomText}
+                                {secondText}
                             </textPath>
                         </text>
-
-                        {secondText && (
-                            <text
-                                fill="#fff"
-                                fontFamily="Montserrat"
-                                fontSize={bottomTextSize}
-                                fontWeight="600"
-                                textAnchor="middle"
-                                style={{ letterSpacing: bottomLetterSpacing }}
-                            >
-                                <textPath
-                                    href={`#${secondId}`}
-                                    startOffset={`${bottomTextOffset}%`}
-                                >
-                                    {secondText}
-                                </textPath>
-                            </text>
-                        )}
-                    </svg>
-                </Box>
-            </Box>
-
-            {/* Mobile: flat description (≤ 600px) */}
-            <Box
-                sx={{
-                    display: "none",
-                    "@media (max-width: 600px)": {
-                        display: "block",
-                        textAlign: "center",
-                        mt: 1
-                    }
-                }}
-            >
-                {secondText && (
-                    <Typography
-                        sx={{
-                            color: "#fff",
-                            fontFamily: "Montserrat",
-                            fontWeight: 600,
-                            fontSize: "14px"
-                        }}
-                    >
-                        {secondText}
-                    </Typography>
-                )}
-                <Typography
-                    sx={{
-                        color: "#fff",
-                        fontFamily: "Montserrat",
-                        fontWeight: 600,
-                        fontSize: "14px"
-                    }}
-                >
-                    {bottomText}
-                </Typography>
+                    )}
+                </svg>
             </Box>
 
             {/* 2. TEXT CONTAINER (Dynamic Height) */}
